major changes
redesigned the system to only expose a "raw" method to the scripting interface, then the scripting user is responsible to construct wrappers to make the process intuitive and editor-friendly. The motivation being that "packed xxx array" data structures are massively more efficient than constructing variant arrays and leveraging them when piercing the interop veil would be more performant.